# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification - Transporte Escolar Eventos
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /alunos:
    get:
      operationId: listAlunos
      tags: [alunos]
      summary: Listar alunos com filtros
      parameters:
        - name: escola_id
          in: query
          schema: { type: integer }
        - name: periodo
          in: query
          schema: { type: string }
        - name: segmento
          in: query
          schema: { type: string }
        - name: turma
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Lista de alunos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Aluno"
    post:
      operationId: createAluno
      tags: [alunos]
      summary: Cadastrar aluno
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlunoInput"
      responses:
        "201":
          description: Aluno criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Aluno"
  /alunos/import:
    post:
      operationId: importAlunos
      tags: [alunos]
      summary: Importar alunos via planilha (base64)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportRequest"
      responses:
        "200":
          description: Resultado da importação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /alunos/contagem:
    get:
      operationId: contagemAlunos
      tags: [alunos]
      summary: Contagem de alunos agrupada por escola, período e turma
      parameters:
        - name: escola_id
          in: query
          schema: { type: integer }
        - name: periodo
          in: query
          schema: { type: string }
        - name: segmento
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Contagens agrupadas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContagemAluno"
  /alunos/{id}:
    get:
      operationId: getAluno
      tags: [alunos]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Aluno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Aluno"
    patch:
      operationId: updateAluno
      tags: [alunos]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlunoUpdate"
      responses:
        "200":
          description: Aluno atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Aluno"
    delete:
      operationId: deleteAluno
      tags: [alunos]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deletado

  # ── Escolas ─────────────────────────────────────────────────────────────────
...
  /empresas:
    get:
      operationId: listEmpresas
      tags: [empresas]
      summary: Listar empresas
      responses:
        "200":
          description: Lista de empresas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Empresa"
    post:
      operationId: createEmpresa
      tags: [empresas]
      summary: Cadastrar empresa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmpresaInput"
      responses:
        "201":
          description: Empresa criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Empresa"
  /empresas/import:
    post:
      operationId: importEmpresas
      tags: [empresas]
      summary: Importar empresas via planilha
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportRequest"
      responses:
        "200":
          description: Resultado da importação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /empresas/{id}:
    get:
      operationId: getEmpresa
      tags: [empresas]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Empresa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Empresa"
    patch:
      operationId: updateEmpresa
      tags: [empresas]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmpresaUpdate"
      responses:
        "200":
          description: Empresa atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Empresa"
    delete:
      operationId: deleteEmpresa
      tags: [empresas]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deletado

  # ── Veículos ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Aluno:
      type: object
      required: [id, nome, ra, escola_id, turma, periodo, segmento]
      properties:
        id:          { type: integer }
        ra:          { type: string }
        nome:        { type: string }
        escola_id:   { type: integer }
        escola_nome: { type: ["string", "null"] }
        turma:       { type: string }
        classe:      { type: ["string", "null"] }
        periodo:     { type: string }
        segmento:    { type: string }
        sexo:        { type: ["string", "null"] }
        nascimento:  { type: ["string", "null"] }
        situacao_matricula: { type: ["string", "null"] }
        zona:        { type: ["string", "null"] }
        frota:       { type: ["string", "null"] }
    AlunoInput:
      type: object
      required: [nome, ra, escola_id, turma, periodo, segmento]
      properties:
        ra:          { type: string }
        nome:        { type: string }
        escola_id:   { type: integer }
        turma:       { type: string }
        classe:      { type: string }
        periodo:     { type: string }
        segmento:    { type: string }
        sexo:        { type: string }
        nascimento:  { type: string }
        situacao_matricula: { type: string }
        zona:        { type: string }
        frota:       { type: string }
    AlunoUpdate:
      type: object
      properties:
        nome:        { type: string }
        turma:       { type: string }
        classe:      { type: string }
        periodo:     { type: string }
        segmento:    { type: string }
        situacao_matricula: { type: string }
        escola_id:   { type: integer }
    ContagemAluno:
      type: object
      required: [escola_id, escola_nome, periodo, segmento, turma, quantidade]
      properties:
        escola_id:   { type: integer }
        escola_nome: { type: string }
        periodo:     { type: string }
        segmento:    { type: string }
        turma:       { type: string }
        quantidade:  { type: integer }

    # ── Escola ────────────────────────────────────────────────────────────────
...
    Empresa:
      type: object
      required: [id, nome]
      properties:
        id:           { type: integer }
        codigo:       { type: ["string", "null"] }
        nome:         { type: string }
        cnpj:         { type: ["string", "null"] }
        responsavel:  { type: ["string", "null"] }
        telefone:     { type: ["string", "null"] }
        email:        { type: ["string", "null"] }
        modalidade:   { type: ["string", "null"] }
        ativo:        { type: boolean }
        total_veiculos: { type: ["integer", "null"] }
    EmpresaInput:
      type: object
      required: [nome]
      properties:
        codigo:      { type: string }
        nome:        { type: string }
        cnpj:        { type: string }
        responsavel: { type: string }
        telefone:    { type: string }
        email:       { type: string }
        modalidade:  { type: string }
        ativo:       { type: boolean }
    EmpresaUpdate:
      type: object
      properties:
        nome:        { type: string }
        cnpj:        { type: string }
        responsavel: { type: string }
        telefone:    { type: string }
        email:       { type: string }
        modalidade:  { type: string }
        ativo:       { type: boolean }

    # ── Veículo ───────────────────────────────────────────────────────────────
...
    ImportRequest:
      type: object
      required: [fileBase64]
      properties:
        fileBase64: { type: string }
    ImportResult:
      type: object
      required: [total, importados, ignorados, erros]
      properties:
        total:     { type: integer }
        importados: { type: integer }
        ignorados:  { type: integer }
        erros:
          type: array
          items:
            type: string