# lib/api-zod/src/generated/types/empresaUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification - Transporte Escolar Eventos
 * OpenAPI spec version: 0.1.0
 */

export interface EmpresaUpdate {
  nome?: string;
  cnpj?: string;
  responsavel?: string;
  telefone?: string;
  email?: string;
  modalidade?: string;
  ativo?: boolean;
}